requirements.txt
flask
pywebview
wmi
requests
psutil